requirements.txt
requests
toml
pillow